+++ examples/with_inline_policy/outputs.tf
@@ -2,6 +2,7 @@
 output "outputs" {
   description = "Example role with inline policy"
   value = [
-    module.inline_policy,
+    module.inline_policy.arn,
+    module.inline_policy.name,
   ]
 }